resetmerge: move tests
